scheduler/requirements.txt
schedule
kafka-python
pymongo
dnspython
